Added welcome message and cleaned up
