--- code/src/components/GameList.js
@@ -1,14 +1,0 @@
-import React from 'react'
-
-import { UserName } from './UserName'
-
-
-export const GameList = () => {
-
-
-  return (
-    <>
-    <UserName />
-    </>
-  )
-}
