+++ code/src/components/WelcomePlayer.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+
+export const WelcomePlayer = ({userName}) => {
+    console.log(userName)
+    return (
+    <div>Welcome {userName}</div>
+    )
+}
+
+
